Data.PolyKinded

Description

Representation of types as constructor + list of types.

Lists of types and application

data LoT k where

LoT k represents a list of types which can be applied to a data type of kind k.

Constructors

LoT0 :: LoT Type
Empty list of types.

(:&&:) :: forall k1 ks. k1 -> LoT ks -> LoT (k1 -> ks) infixr 5
Cons a type with a list of types.

type family (f :: k) :@@: (tys :: LoT k) where ...

Apply a list of types to a type constructor.

>>> :kind! Either :@@: (Int :&&: Bool :&&: LoT0)
Either :@@: (Int :&&: Bool :&&: LoT0) :: Type
= Either Int Bool

Equations

(f :: Type) :@@: (_1 :: LoT Type) = f

(f :: k -> k') :@@: (as :: LoT (k -> k')) = f (HeadLoT as) :@@: TailLoT as

type LoT1 (a :: k) = a ':&&: 'LoT0

List of types with a single element.

type LoT2 (a :: k) (b :: k1) = a ':&&: (b ':&&: 'LoT0)

List of types with two elements.

type family HeadLoT (tys :: LoT (k -> k')) :: k where ...

Head of a non-empty list of types.

>>> :kind! HeadLoT (Int :&&: LoT0)
HeadLoT (Int :&&: LoT0) :: Type
= Int

Equations

HeadLoT (a ':&&: _1 :: LoT (k -> k')) = a

type family TailLoT (tys :: LoT (k -> k')) :: LoT k' where ...

Tail of a non-empty list of types.

>>> :kind! TailLoT (Int :&&: Bool :&&: LoT0)
TailLoT (Int :&&: Bool :&&: LoT0) :: LoT (Type -> Type)
= Bool :&&: LoT0

Equations

TailLoT (_1 ':&&: as :: LoT (k -> k')) = as

type family SpineLoT (tys :: LoT k) = (tys' :: LoT k) | tys' -> k tys where ...

Construct the spine of a list of types whose length is known.

It can be useful to introduce unification variables for lists of types which will be fully instantiated during constraint resolution. A constraint p ~ SpineLoT p will thus instantiate the spine of p.

On concrete lists, this is the identity function.

Equations

SpineLoT (a ':&&: as :: LoT (k -> ks)) = a ':&&: SpineLoT as

SpineLoT 'LoT0 = 'LoT0

Singleton for list of types

data SLoT (l :: LoT k) where

Constructors

SLoT0 :: SLoT 'LoT0

SLoTA :: forall {k1} {k2} (t :: k1) (ts :: LoT k2). Proxy t -> SLoT ts -> SLoT (t ':&&: ts)

class SForLoT (l :: LoT k) where

Methods

slot :: SLoT l

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

Splitting types

type SplitF (t :: d) (f :: k) = SplitF' t f 'LoT0

Split a type t until the constructor f is found.

>>> :kind! SplitF (Either Int Bool) Either
SplitF (Either Int Bool) Either :: LoT (Type -> Type -> Type)
= Int :&&: (Bool :&&: LoT0)
>>> :kind! SplitF (Either Int Bool) (Either Int)
SplitF (Either Int Bool) (Either Int) :: LoT (Type -> Type)
= Bool :&&: LoT0

data Nat

Simple natural numbers.

Constructors

Z

S Nat

data TyEnv where

A type constructor and a list of types that can be applied to it.

Constructors

TyEnv :: forall k. k -> LoT k -> TyEnv

type family SplitN (n :: Nat) t :: TyEnv where ...

Split a type t until its list of types has length n.

>>> :kind! SplitN (S (S Z)) (Either Int Bool)
SplitN (S (S Z)) (Either Int Bool) :: TyEnv
= 'TyEnv Either (Int :&&: (Bool :&&: LoT0))
>>> :kind! SplitN (S Z) (Either Int Bool)
SplitN (S Z) (Either Int Bool) :: TyEnv
= 'TyEnv (Either Int) (Bool :&&: LoT0)

Equations

SplitN n t = SplitN' n t 'LoT0